# Refactor bytestream proxy to use byte_stream_server.Write