I did this to make it easier to optimize `byte_stream_server.Write` without having to also handle the `byte_stream_server.BeginWrite` path.

I think the new version is pretty clean, but a little bit harder to follow. In both the `Read` and `Write` paths, I'm creating an object that can be passed to `byte_stream_server.Write`, and each time `Recv()` is called on it, it also does some work on the side (forwarding to the client in Read, and writing to remote in Write).

Somehow, this the proxy faster. Benchmark results with new benchmarks from https://github.com/buildbuddy-io/buildbuddy/pull/9599:
```
                                             │     base      │               refactor               │
                                             │    sec/op     │    sec/op     vs base                │
ReadAlwaysPresent-24                            63.13µ ± 17%   65.96µ ±  9%        ~ (p=0.606 n=11)
ReadThroughCache/zstd=false/size=10000-24       143.6µ ±  2%   145.5µ ±  3%        ~ (p=0.949 n=11)
ReadThroughCache/zstd=false/size=3000000-24     4.241m ±  6%   3.112m ±  6%  -26.61% (p=0.000 n=11)
ReadThroughCache/zstd=false/size=20000000-24    23.17m ± 10%   20.62m ±  7%  -10.97% (p=0.016 n=11)
ReadThroughCache/zstd=true/size=10000-24        240.3µ ±  2%   246.3µ ±  2%   +2.49% (p=0.023 n=11)
ReadThroughCache/zstd=true/size=3000000-24      12.91m ±  2%   12.57m ±  2%   -2.63% (p=0.004 n=11)
ReadThroughCache/zstd=true/size=20000000-24     64.66m ±  3%   60.15m ±  4%   -6.98% (p=0.004 n=11)
WriteAlreadyExists-24                           118.2µ ±  4%   121.2µ ±  8%        ~ (p=0.401 n=11)
WriteUnique/zstd=false/size=10000-24            143.3µ ±  5%   141.6µ ±  7%        ~ (p=0.478 n=11)
WriteUnique/zstd=false/size=3000000-24          5.783m ± 20%   4.213m ±  3%  -27.15% (p=0.000 n=11)
WriteUnique/zstd=false/size=20000000-24         22.17m ± 11%   21.04m ±  7%   -5.09% (p=0.019 n=11)
WriteUnique/zstd=true/size=10000-24             178.2µ ±  1%   164.8µ ±  2%   -7.50% (p=0.000 n=11)
WriteUnique/zstd=true/size=3000000-24          10.010m ± 25%   4.536m ± 18%  -54.68% (p=0.000 n=11)
WriteUnique/zstd=true/size=20000000-24          55.98m ± 10%   38.79m ± 11%  -30.70% (p=0.000 n=11)
geomean                                         2.127m         1.839m        -13.53%

                                             │     base     │              refactor               │
                                             │     B/op     │     B/op      vs base               │
ReadAlwaysPresent-24                           111.1Ki ± 1%   111.0Ki ± 1%       ~ (p=0.898 n=11)
ReadThroughCache/zstd=false/size=10000-24      98.75Ki ± 0%   98.82Ki ± 0%       ~ (p=0.401 n=11)
ReadThroughCache/zstd=false/size=3000000-24    11.65Mi ± 0%   11.68Mi ± 0%       ~ (p=0.101 n=11)
ReadThroughCache/zstd=false/size=20000000-24   115.7Mi ± 0%   115.1Mi ± 0%       ~ (p=0.065 n=11)
ReadThroughCache/zstd=true/size=10000-24       78.88Ki ± 0%   79.08Ki ± 0%  +0.25% (p=0.000 n=11)
ReadThroughCache/zstd=true/size=3000000-24     10.95Mi ± 1%   10.82Mi ± 1%       ~ (p=0.151 n=11)
ReadThroughCache/zstd=true/size=20000000-24    91.64Mi ± 9%   90.76Mi ± 2%       ~ (p=0.193 n=11)
WriteAlreadyExists-24                          134.5Ki ± 0%   137.1Ki ± 0%  +1.91% (p=0.000 n=11)
WriteUnique/zstd=false/size=10000-24           96.20Ki ± 0%   96.29Ki ± 0%       ~ (p=0.133 n=11)
WriteUnique/zstd=false/size=3000000-24         11.60Mi ± 0%   11.62Mi ± 0%       ~ (p=0.606 n=11)
WriteUnique/zstd=false/size=20000000-24        110.5Mi ± 3%   110.1Mi ± 1%       ~ (p=0.699 n=11)
WriteUnique/zstd=true/size=10000-24            83.31Ki ± 0%   83.46Ki ± 0%  +0.18% (p=0.000 n=11)
WriteUnique/zstd=true/size=3000000-24          17.85Mi ± 1%   17.72Mi ± 0%       ~ (p=0.065 n=11)
WriteUnique/zstd=true/size=20000000-24         148.2Mi ± 3%   146.8Mi ± 3%       ~ (p=0.748 n=11)
geomean                                        2.946Mi        2.942Mi       -0.14%

                                             │    base     │              refactor              │
                                             │  allocs/op  │  allocs/op   vs base               │
ReadAlwaysPresent-24                            418.0 ± 0%    418.0 ± 0%       ~ (p=1.000 n=11)
ReadThroughCache/zstd=false/size=10000-24       867.0 ± 0%    868.0 ± 0%  +0.12% (p=0.000 n=11)
ReadThroughCache/zstd=false/size=3000000-24    1.053k ± 2%   1.061k ± 1%       ~ (p=0.076 n=11)
ReadThroughCache/zstd=false/size=20000000-24   2.271k ± 3%   2.294k ± 0%  +1.01% (p=0.000 n=11)
ReadThroughCache/zstd=true/size=10000-24        881.0 ± 0%    889.0 ± 0%  +0.91% (p=0.000 n=11)
ReadThroughCache/zstd=true/size=3000000-24      956.0 ± 0%    977.0 ± 0%  +2.20% (p=0.000 n=11)
ReadThroughCache/zstd=true/size=20000000-24    1.354k ± 1%   1.384k ± 1%  +2.22% (p=0.000 n=11)
WriteAlreadyExists-24                           838.0 ± 0%    861.0 ± 0%  +2.74% (p=0.000 n=11)
WriteUnique/zstd=false/size=10000-24            837.0 ± 0%    839.0 ± 0%  +0.24% (p=0.000 n=11)
WriteUnique/zstd=false/size=3000000-24         1.011k ± 0%   1.069k ± 0%  +5.74% (p=0.000 n=11)
WriteUnique/zstd=false/size=20000000-24        2.005k ± 2%   2.132k ± 1%  +6.33% (p=0.000 n=11)
WriteUnique/zstd=true/size=10000-24             846.0 ± 0%    850.0 ± 0%  +0.47% (p=0.000 n=11)
WriteUnique/zstd=true/size=3000000-24           925.0 ± 0%    942.0 ± 0%  +1.84% (p=0.000 n=11)
WriteUnique/zstd=true/size=20000000-24         1.325k ± 1%   1.348k ± 1%  +1.74% (p=0.001 n=11)
geomean                                        1.027k        1.046k       +1.86%
```