@vanja-p 2025-06-10
Using the grpc interface again.